refactor: DemoController에서 사용하는 함수는 DemoService로 이동
